Split Tests into two Test Suites
